requirements.txt
pandas
matplotlib.pyplot
streamlit
numpy
openpyxl